removed padding file + option to use padding in makefile
Previously, building on windows without unix commands like dd was possible using a padding file and the copy command.
This was clunky and didn't always work properly, as the output image varied in size depending
on the size of the built OS. I removed this because this OS should be built on windows
using something akin to MSYS2 with built-in unix commands.